Fix MW/mW derivation bug
Fixes #27.

The cause were missing qudt:conversionMultiplier triples for unit:MW and unit:mW. The upstream fix is merged and will have an effect for us when qudt is released and we upgrade. Meanwhile, the triples are added.

We also upgrade to qudt v2.1.20, which causes a minor release for us.